working on CRUD for games and publisher
TODO:
- update PUT and PATCH following Devtiro example in crash course
- test all endpoints for games and publishers

+++ src/main/java/com/barux/e4cSpring/publisher/PublisherRepository.java
@@ -3,5 +3,5 @@ package com.barux.e4cSpring.publisher;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface PublisherRepository extends JpaRepository<Publisher, Long>, PagingAndSortingRepository<Publisher, Long> {
+public interface PublisherRepository extends JpaRepository<Publisher, Integer>, PagingAndSortingRepository<Publisher, Integer> {
 }
